pyproject.toml
[project]
name = "reg-gan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.9",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "scikit-image>=0.26.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "wandb>=0.26.1",
]
